+++ packages/component-utilities/src/getCompletedData.js
@@ -25,11 +25,48 @@ export default function getCompletedData(data, x, y, series, nullsZero = false, 
 		return a;
 	}, {});
 
-	const output = [];
-	for (const value of Object.values(groups)) {
-		let xIsDate = value[0]?.[x] instanceof Date;
+	// Ensures that all permutations of this map exist in the output
+	// e.g. can include series and x values to ensure that all series have all x values
+	const expandKeys = {};
 
-		const nullySpec = { series: null };
+	const xIsDate = data[0]?.[x] instanceof Date;
+
+	/** @type {Array<number | string>} */
+	let xDistinct;
+	const exampleX = data[0]?.[x];
+
+	switch (typeof exampleX) {
+		case 'object':
+			// If x is not a date; this shouldn't be hit, abort!
+			if (!(exampleX instanceof Date)) {
+				throw new Error('Unexpected object property, expected string, date, or number');
+			}
+			// Map dates to numeric values
+			xDistinct = getDistinctValues(
+				data.map((d) => ({ [x]: d[x].getTime() })),
+				x
+			);
+			// We don't fillX here because date numbers are very large, so a small interval would create a _massive_ array
+			break;
+		case 'number':
+			// Numbers are the most straightforward
+			xDistinct = getDistinctValues(data, x);
+			if (fillX) {
+				// Attempt to derive the interval between X values and interpolate missing values in that set (within the bounds of min/max)
+				const interval = findInterval(xDistinct);
+				expandKeys[x] = vectorSeq(xDistinct, interval);
+			}
+			break;
+		case 'string':
+			xDistinct = getDistinctValues(data, x);
+			expandKeys[x] = xDistinct;
+			break;
+	}
+
+	const output = [];
+
+	for (const value of Object.values(groups)) {
+		const nullySpec = series ? { [series]: null } : {};
 		if (nullsZero) {
 			nullySpec[y] = 0;
 		} else {
@@ -37,44 +74,27 @@ export default function getCompletedData(data, x, y, series, nullsZero = false, 
 			nullySpec[y] = null;
 		}
 
-		const expandKeys = {};
-		if (fillX) {
-			/** @type {Array<number>} */
-			let xDistinct;
-
-			if (xIsDate)
-				xDistinct = getDistinctValues(
-					value.map((d) => ({ [x]: d[x].getTime() })),
-					x
-				);
-			else xDistinct = getDistinctValues(value, x);
-
-			/** @type {number} */
-			let interval = findInterval(xDistinct);
-
-			// Array of all possible x values
-			expandKeys[x] = vectorSeq(xDistinct, interval);
-		}
 		if (series) {
 			expandKeys[series] = series;
 		}
 
 		const tidyFuncs = [];
-
 		if (Object.keys(expandKeys).length === 0) {
-			tidyFuncs.push(complete([x], nullySpec));
 			// empty object, no special configuration
+			tidyFuncs.push(complete([x], nullySpec));
 		} else {
 			tidyFuncs.push(complete(expandKeys, nullySpec));
 		}
 		if (xIsDate) {
+			// Ensure that x is actually a date
 			tidyFuncs.push(
 				mutate({
 					[x]: (val) => new Date(val[x])
 				})
 			);
 		}
-		output.push(...tidy(value, ...tidyFuncs));
+
+		output.push(tidy(value, ...tidyFuncs));
 	}
-	return output;
+	return output.flat();
 }